+++ clients/fides-js/src/lib/gpp/modules/cmpapi/lib/cjs/encoder/field/TcfCaV1Field.d.ts
@@ -1,0 +1,29 @@
+export declare enum TcfCaV1Field {
+    VERSION = "Version",
+    CREATED = "Created",
+    LAST_UPDATED = "LastUpdated",
+    CMP_ID = "CmpId",
+    CMP_VERSION = "CmpVersion",
+    CONSENT_SCREEN = "ConsentScreen",
+    CONSENT_LANGUAGE = "ConsentLanguage",
+    VENDOR_LIST_VERSION = "VendorListVersion",
+    TCF_POLICY_VERSION = "TcfPolicyVersion",
+    USE_NON_STANDARD_STACKS = "UseNonStandardStacks",
+    SPECIAL_FEATURE_EXPRESS_CONSENT = "SpecialFeatureExpressConsent",
+    PUB_PURPOSES_SEGMENT_TYPE = "PubPurposesSegmentType",
+    PURPOSES_EXPRESS_CONSENT = "PurposesExpressConsent",
+    PURPOSES_IMPLIED_CONSENT = "PurposesImpliedConsent",
+    VENDOR_EXPRESS_CONSENT = "VendorExpressConsent",
+    VENDOR_IMPLIED_CONSENT = "VendorImpliedConsent",
+    PUB_RESTRICTIONS = "PubRestrictions",
+    PUB_PURPOSES_EXPRESS_CONSENT = "PubPurposesExpressConsent",
+    PUB_PURPOSES_IMPLIED_CONSENT = "PubPurposesImpliedConsent",
+    NUM_CUSTOM_PURPOSES = "NumCustomPurposes",
+    CUSTOM_PURPOSES_EXPRESS_CONSENT = "CustomPurposesExpressConsent",
+    CUSTOM_PURPOSES_IMPLIED_CONSENT = "CustomPurposesImpliedConsent",
+    DISCLOSED_VENDORS_SEGMENT_TYPE = "DisclosedVendorsSegmentType",
+    DISCLOSED_VENDORS = "DisclosedVendors"
+}
+export declare const TCFCAV1_CORE_SEGMENT_FIELD_NAMES: TcfCaV1Field[];
+export declare const TCFCAV1_PUBLISHER_PURPOSES_SEGMENT_FIELD_NAMES: TcfCaV1Field[];
+export declare const TCFCAV1_DISCLOSED_VENDORS_SEGMENT_FIELD_NAMES: TcfCaV1Field[];