+++ clients/fides-js/src/lib/gpp/modules/cmpapi/lib/mjs/index.js
@@ -1,0 +1,5 @@
+export * from "./cmpapi/index.js";
+export * from "./encoder/index.js";
+export * from "./gvl/index.js";
+export * from "./CmpApi.js";
+export * from "./GVL.js";